requirements.txt
Django>=4.2
notion-client
python-dotenv
